refac(zk): move `zk/base/Halo2XXX` classes to `zk/base/halo2`
