+++ tachyon/zk/base/BUILD.bazel
@@ -24,41 +24,6 @@ tachyon_cc_library(
 )
 
 tachyon_cc_library(
-    name = "halo2_prover",
-    hdrs = ["halo2_prover.h"],
-    deps = [
-        ":halo2_random_field_generator",
-        "//tachyon/zk/base/entities:prover",
-    ],
-)
-
-tachyon_cc_library(
-    name = "halo2_prover_test",
-    testonly = True,
-    hdrs = ["halo2_prover_test.h"],
-    deps = [
-        ":blinder",
-        ":halo2_prover",
-        "//tachyon/math/elliptic_curves/bn/bn254:g1",
-        "//tachyon/math/elliptic_curves/bn/bn254:g2",
-        "//tachyon/math/polynomials/univariate:univariate_evaluation_domain_factory",
-        "//tachyon/zk/base/commitments:kzg_commitment_scheme_extension",
-        "//tachyon/zk/transcript:poseidon_transcript",
-        "@com_google_absl//absl/memory",
-    ],
-)
-
-tachyon_cc_library(
-    name = "halo2_random_field_generator",
-    hdrs = ["halo2_random_field_generator.h"],
-    deps = [
-        "//tachyon/crypto/random/xor_shift:xor_shift_rng",
-        "//tachyon/math/base:big_int",
-        "@com_google_googletest//:gtest_prod",
-    ],
-)
-
-tachyon_cc_library(
     name = "point_stringifier",
     hdrs = ["point_stringifier.h"],
     deps = [
@@ -113,18 +78,15 @@ tachyon_cc_unittest(
     name = "base_unittests",
     srcs = [
         "blinder_unittest.cc",
-        "halo2_random_field_generator_unittest.cc",
         "ref_unittest.cc",
         "value_unittest.cc",
     ],
     deps = [
         ":blinder",
-        ":halo2_random_field_generator",
         ":ref",
         ":value",
         "//tachyon/base:random",
         "//tachyon/base/containers:container_util",
-        "//tachyon/math/elliptic_curves/bn/bn254:g1",
         "//tachyon/math/finite_fields/test:gf7",
     ],
 )
