+++ tachyon/zk/base/entities/prover_unittest.cc
@@ -1,6 +1,6 @@
 #include "gtest/gtest.h"
 
-#include "tachyon/zk/base/halo2_prover_test.h"
+#include "tachyon/zk/base/halo2/halo2_prover_test.h"
 
 namespace tachyon::zk {
