+++ tachyon/zk/base/halo2/BUILD.bazel
@@ -1,0 +1,48 @@
+load("//bazel:tachyon_cc.bzl", "tachyon_cc_library", "tachyon_cc_unittest")
+
+package(default_visibility = ["//visibility:public"])
+
+tachyon_cc_library(
+    name = "halo2_prover",
+    hdrs = ["halo2_prover.h"],
+    deps = [
+        ":halo2_random_field_generator",
+        "//tachyon/zk/base/entities:prover",
+    ],
+)
+
+tachyon_cc_library(
+    name = "halo2_prover_test",
+    testonly = True,
+    hdrs = ["halo2_prover_test.h"],
+    deps = [
+        ":halo2_prover",
+        "//tachyon/math/elliptic_curves/bn/bn254:g1",
+        "//tachyon/math/elliptic_curves/bn/bn254:g2",
+        "//tachyon/math/polynomials/univariate:univariate_evaluation_domain_factory",
+        "//tachyon/zk/base:blinder",
+        "//tachyon/zk/base/commitments:kzg_commitment_scheme_extension",
+        "//tachyon/zk/transcript:poseidon_transcript",
+        "@com_google_absl//absl/memory",
+    ],
+)
+
+tachyon_cc_library(
+    name = "halo2_random_field_generator",
+    hdrs = ["halo2_random_field_generator.h"],
+    deps = [
+        "//tachyon/crypto/random/xor_shift:xor_shift_rng",
+        "//tachyon/math/base:big_int",
+        "//tachyon/zk/base:random_field_generator",
+        "@com_google_googletest//:gtest_prod",
+    ],
+)
+
+tachyon_cc_unittest(
+    name = "halo2_unittests",
+    srcs = ["halo2_random_field_generator_unittest.cc"],
+    deps = [
+        ":halo2_random_field_generator",
+        "//tachyon/math/elliptic_curves/bn/bn254:fq",
+    ],
+)
