+++ tachyon/zk/base/halo2/halo2_random_field_generator.h
@@ -1,5 +1,5 @@
-#ifndef TACHYON_ZK_BASE_HALO2_RANDOM_FIELD_GENERATOR_H_
-#define TACHYON_ZK_BASE_HALO2_RANDOM_FIELD_GENERATOR_H_
+#ifndef TACHYON_ZK_BASE_HALO2_HALO2_RANDOM_FIELD_GENERATOR_H_
+#define TACHYON_ZK_BASE_HALO2_HALO2_RANDOM_FIELD_GENERATOR_H_
 
 #include "gtest/gtest_prod.h"
 
@@ -62,4 +62,4 @@ class Halo2RandomFieldGenerator : public RandomFieldGenerator<F> {
 
 }  // namespace tachyon::zk
 
-#endif  // TACHYON_ZK_BASE_HALO2_RANDOM_FIELD_GENERATOR_H_
+#endif  // TACHYON_ZK_BASE_HALO2_HALO2_RANDOM_FIELD_GENERATOR_H_
