+++ tachyon/zk/plonk/keys/verifying_key_unittest.cc
@@ -8,7 +8,7 @@
 
 #include "gtest/gtest.h"
 
-#include "tachyon/zk/base/halo2_prover_test.h"
+#include "tachyon/zk/base/halo2/halo2_prover_test.h"
 #include "tachyon/zk/plonk/keys/halo2/pinned_verifying_key.h"
 
 namespace tachyon::zk {
