+++ tachyon/zk/plonk/permutation/BUILD.bazel
@@ -170,6 +170,6 @@ tachyon_cc_unittest(
         ":permutation_assembly",
         ":permutation_table_store",
         "//tachyon/base/buffer:vector_buffer",
-        "//tachyon/zk/base:halo2_prover_test",
+        "//tachyon/zk/base/halo2:halo2_prover_test",
     ],
 )
